ロジックをwrapper層へ移動
